c++/coroutines: fix actor cases not being added to the current switch [PR109867]
Previously, we were building and inserting case_labels manually, which
led to them not being added into the currently running switch via
c_add_case_label.  This led to false diagnostics that the user could not
act on.

The use of temp_override on current_function_decl is a temporary hack,
as we intend to factor out generating the outlined functions (including
the actor) to happen later in finish_function, when we can create a new
function using the standard routines for that, similar to
finish_function_contracts.

PR c++/109867 - -Wswitch-default reports missing default in coroutine

gcc/cp/ChangeLog:

	PR c++/109867
	* coroutines.cc (expand_one_await_expression): Replace uses of
	build_case_label with finish_case_label.
	(build_actor_fn): Ditto.  Also adjust current_function_decl.
	(create_anon_label_with_ctx): Remove now-unused function.

gcc/testsuite/ChangeLog:

	PR c++/109867
	* g++.dg/coroutines/torture/pr109867.C: New test.