c++, coroutines: Split the ramp build into a separate function.
This is primarily preparation to partition the functionality of the
coroutine transform into analysis, ramp generation and then (later)
synthesis of the coroutine body.  The patch does fix one latent
issue in the ordering of DTORs for frame parameter copies (to ensure
that they are processed in reverse order to the copy creation).

gcc/cp/ChangeLog:

	* coroutines.cc (build_actor_fn): Arrange to apply any
	required parameter copy DTORs in reverse order to their
	creation.
	(morph_fn_to_coro): Split the ramp function completion
	into a separate function.
	(build_ramp_function): New.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>