c++, coroutines: Separate allocator work from the ramp body build.
This splits out the building of the allocation and deallocation expressions
and runs them early in the ramp build, so that we can exit if they are not
usable, before we start building the ramp body.

Likewise move checks for other required resources to the begining of the
ramp builder.

This is preparation for work needed to update the allocation/destruction
in cases where we have excess alignment of the promise or other saved frame
state.

gcc/cp/ChangeLog:

	* coroutines.cc (coro_get_frame_dtor): Rename...
	(build_coroutine_frame_delete_expr):... to this.
	(build_actor_fn): Use revised frame delete function.
	(build_coroutine_frame_alloc_expr): New.
	(cp_coroutine_transform::complete_ramp_function): Rename...
	(cp_coroutine_transform::build_ramp_function): ... to this.
	Reorder code to carry out checks for prerequisites before the
	codegen. Split out the allocation/delete code.
	(cp_coroutine_transform::apply_transforms): Use revised name.
	* coroutines.h: Rename function.

gcc/testsuite/ChangeLog:

	* g++.dg/coroutines/coro-bad-gro-00-class-gro-scalar-return.C:
	Use revised diagnostics.
	* g++.dg/coroutines/coro-bad-gro-01-void-gro-non-class-coro.C:
	Likewise.
	* g++.dg/coroutines/coro-bad-grooaf-00-static.C: Likewise.
	* g++.dg/coroutines/ramp-return-b.C: Likewise.

Signed-off-by: Iain Sandoe <iain@sandoe.co.uk>